+++ py/LGA_MediaManager_logging.py
@@ -1,7 +1,7 @@
 """
 _______________________________________
 
-  LGA_MediaManager_logging v2.42 | Lega
+  LGA_MediaManager_logging v2.43 | Lega
   Logger compartido del Media Manager
 
   v2.25: Se le pone header con version, para que acompane al resto de